tool/typora_md_to_km/requirements.txt
jinja2==2.10.1
configparser==3.7.4
PyYAML==5.1 